// src/redux-toolkit/store/users/type.ts
export interface IUser {
  id: number;
  name: string;
  username: string;
  email: string;
  address: {
    street: string;
    suite: string;
    city: string;
    zipcode: string;
    geo: {
      lat: string;
      lng: string;
    };
  };
}

export interface IState{
  loading: boolean,
  error: boolean,
  users: IUser[]
}